Minor bug/behavior fixes.
Hand still ends and goes to payouts ignoring any remaining hands the player has due to a split. The problem is in BlackjackModel.cs Hit(). If the player goes bust after a card is dealt the game immediately jumps to resolve/payout the hand. This needs to go through the player instead--verify if the hand is truly done first.

+++ Blackjack/Controller/BlackjackController.cs
@@ -32,7 +32,7 @@ namespace Blackjack.Application
                 case "stand":
                     model.Stand();
                     break;
-                case "decimal":
+                case "double":
                     model.DoubleDown();
                     break;
                 case "split":
